refactoring centos7

+++ centos7/mysql56.sh
@@ -5,20 +5,14 @@ source prepare.sh
 #<<<<<<<<<<
 
 
-# root only
-if [ ${MYUSER} != "root" ]
-then
-  echo "${MYUSER} can not run ${MYNAME}"
-  exit 1
-fi
+i
+[ $(isroot) ] || (echo "${MYUSER} can not run ${MYNAME}"; exit 1)
 
 # repository package
+pushd ${PRVENV_INSTALL_WORK_DIR}
+
 declare -r REPO_RPM=mysql-community-release-el7-5.noarch.rpm
-cd /tmp
-if [ -f ${REPO_RPM} ]
-then
-  rm -f ${REPO_RPM}
-fi
+[ -f ${REPO_RPM} ] && rm -f ${REPO_RPM}
 ${WGETCMD} http://dev.mysql.com/get/${REPO_RPM}
 
 # install repository package
