+++ centos7/php5-fpm.sh
@@ -5,12 +5,7 @@ source prepare.sh
 #<<<<<<<<<<
 
 
-# root only
-if [ ${MYUSER} != "root" ]
-then
-  echo "${MYUSER} can not run ${MYNAME}"
-  exit 1
-fi
+[ $(isroot) ] || (echo "${MYUSER} can not run ${MYNAME}"; exit 1)
 
 # install
 ${PRVENV_CMD_PKG_INS} php56-fpm
@@ -20,10 +15,8 @@ ${PRVENV_CMD_PKG_INS} php56-cli php56-gd php56-mysqlnd php56-xmlrpc
 
 # modify php.ini
 declare -r PHP_INI_FILE=/etc/php56/fpm/php.ini
-if [ ! -f ${PHP_INI_FILE}.org ]
-then
-  cp ${PHP_INI_FILE} ${PHP_INI_FILE}.org
-fi
+[ -f ${PHP_INI_FILE}.org ] || cp ${PHP_INI_FILE} ${PHP_INI_FILE}.org
+
 ## fix_pathinfo=1 is insecure
 sed -i "s/;cgi.fix_pathinfo=1/cgi.fix_pathinfo=0/g" ${PHP_INI_FILE}
